+++ src/main/java/org/folio/services/caches/ConsortiumDataCache.java
@@ -35,7 +35,7 @@ public class ConsortiumDataCache {
   private static final String EXPIRATION_TIME_PARAM = "cache.consortium-data.expiration.time.seconds";
   private static final String DEFAULT_EXPIRATION_TIME_SECONDS = "300";
   private static final String USER_TENANTS_PATH = "/user-tenants?limit=1"; //NOSONAR
-  private static final String CONSORTIUM_TENANTS_PATH = "/consortia/%s/tenants";
+  private static final String CONSORTIUM_TENANTS_PATH = "/consortia/%s/tenants"; //NOSONAR
   private static final String USER_TENANTS_FIELD = "userTenants";
   private static final String CONSORTIUM_TENANTS_FIELD = "tenants";
   private static final String CENTRAL_TENANT_ID_FIELD = "centralTenantId";
@@ -111,14 +111,12 @@ public class ConsortiumDataCache {
                                                      Map<String, String> headers) {
     var request = getHttpRequest(headers, CONSORTIUM_TENANTS_PATH.formatted(consortiumId));
     return getResponse(tenantId, request)
-      .map(responseBody -> {
-          return responseBody.map(entries -> entries.getJsonArray(CONSORTIUM_TENANTS_FIELD)
-            .stream()
-            .map(o -> ((JsonObject) o).mapTo(ConsortiumTenant.class))
-            .filter(consortiumTenant -> !consortiumTenant.isCentral())
-            .map(ConsortiumTenant::id)
-            .toList()).orElse(Collections.emptyList());
-        }
+      .map(responseBody -> responseBody.map(entries -> entries.getJsonArray(CONSORTIUM_TENANTS_FIELD)
+        .stream()
+        .map(o -> ((JsonObject) o).mapTo(ConsortiumTenant.class))
+        .filter(consortiumTenant -> !consortiumTenant.isCentral())
+        .map(ConsortiumTenant::id)
+        .toList()).orElse(Collections.emptyList())
       )
       .recover(throwable -> succeededFuture(Collections.emptyList()));
   }