+++ src/main/java/org/folio/services/consortium/SynchronizationEventProcessor.java
@@ -3,7 +3,7 @@ package org.folio.services.consortium;
 import io.vertx.core.Future;
 import org.folio.services.domainevent.DomainEvent;
 
-public interface SynchronizationEventProcessor<T> {
+public interface SynchronizationEventProcessor {
 
   Future<String> process(DomainEvent<?> domainEvent, String key, SynchronizationContext synchronizationContext);
 }